+++ weaver/datatype.py
@@ -1929,6 +1929,19 @@ class Process(Base):
         return self.type != ProcessType.BUILTIN
 
     @property
+    def deployment_profile(self):
+        # type: () -> str
+        base = "http://www.opengis.net/profiles/eoc/"
+        _typ = self.type
+        if _typ == ProcessType.APPLICATION:
+            profile = base + "dockerizedApplication"
+        elif "wps" in _typ:
+            profile = base + "wpsApplication"
+        else:
+            profile = base + _typ
+        return profile
+
+    @property
     def package(self):
         # type: () -> Optional[CWL]
         """
@@ -2163,7 +2176,10 @@ class Process(Base):
         """
         process = self.dict()
         links = self.links()
-        process.update({"links": links})
+        process.update({
+            "deploymentProfile": self.deployment_profile,
+            "links": links
+        })
         # force selection of schema to avoid ambiguity
         if str(schema or ProcessSchema.OGC).upper() == ProcessSchema.OLD:
             # nested process fields + I/O as lists